limit warnings

+++ ext/obj_ext/RIGSCore.h
@@ -37,22 +37,22 @@
 void  rb_objc_release(id objc_object);
 VALUE rb_objc_new(int rigs_argc, VALUE *rigs_argv, VALUE rb_class);
 
-BOOL rb_objc_convert_to_objc(VALUE rb_val, void **data, int offset, const char *type);
-BOOL rb_objc_convert_to_rb(void *data, int offset, const char *type, VALUE *rb_val_ptr, BOOL autoconvert);
+BOOL rb_objc_convert_to_objc(VALUE rb_val, void **data, size_t offset, const char *type);
+BOOL rb_objc_convert_to_rb(void *data, size_t offset, const char *type, VALUE *rb_val_ptr, BOOL autoconvert);
 
 VALUE rb_objc_send(int rigs_argc, VALUE *rigs_argv, VALUE rb_self);
 VALUE rb_objc_invoke(int rigs_argc, VALUE *rigs_argv, VALUE rb_self);
 
 unsigned int rb_objc_register_instance_methods(Class objc_class, VALUE rb_class);
 unsigned int rb_objc_register_class_methods(Class objc_class, VALUE rb_class);
-VALUE rb_objc_register_class_from_objc (Class objc_class);
+VALUE rb_objc_register_class_from_objc(Class objc_class);
 VALUE rb_objc_register_class_from_ruby(VALUE rb_self, VALUE rb_name);
 
 void rb_objc_register_float_from_objc(const char *name, double value);
 void rb_objc_register_integer_from_objc(const char *name, long long value);
-void rb_objc_register_struct_from_objc(const char *key, const char *name, const char *args[], int argCount);
-void rb_objc_register_format_string_from_objc(const char *selector, int index);
-void rb_objc_register_block_from_objc(const char *selector, int index, const char *objcTypes);
+void rb_objc_register_struct_from_objc(const char *key, const char *name, const char *args[], size_t argCount);
+void rb_objc_register_format_string_from_objc(const char *selector, size_t index);
+void rb_objc_register_block_from_objc(const char *selector, size_t index, const char *objcTypes);
 void rb_objc_register_constant_from_objc(const char *name, const char *type);
 void rb_objc_register_function_from_objc(const char *name, const char *objcTypes);
 void rb_objc_register_protocol_from_objc(const char *protocolName);
